backend/requirements.txt
fastapi==0.115.6
fastapi-cli==0.0.7
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
keras==3.7.0
mysql-connector-python==9.1.0
numpy==2.0.2
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-multipart==0.0.19
requests==2.32.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
